+++ PoliTechAPI/pt-launcher/src/main/java/ru/pt/api/admin/CommissionController.java
@@ -37,7 +37,7 @@ public class CommissionController  {
     public ResponseEntity<List<CommissionRateDto>> getConfigurations(
             @PathVariable String tenantCode,
             @RequestParam(required = true) Long accountId,
-            @RequestParam(required = false) Integer productId,
+            @RequestParam(required = false) Long productId,
             @RequestParam(required = false) String action) {
         List<CommissionRateDto> list = commissionService.getConfigurations(accountId, productId, CommissionAction.fromValue(action));
         return ResponseEntity.ok(list);
